move banner js
